MPS when daily optimization : extracting optimization period as a string in .h/.cpp files
